finance overlay broken into components
