+++ packages/desktop/src/components/finance/components/deployDomain/deployDomain.js
@@ -1,0 +1,151 @@
+import React, { useState } from 'react'
+import { useSelector, useDispatch } from 'react-redux';
+import { Button } from 'antd';
+import { Input } from 'antd';
+import { getWalletProvider, getWallet, deployDomainDAO } from '../../../../service/contracts/shared';
+
+let registryAddress = "0x132efA3675cd66aA2780e01C095e2337188b0F6b";
+
+const DeployDomain = () => {
+  // need to know if this current domain has been registered
+  // if it is deployed and if you have invested
+  // render out diff stuff depending on the above
+  const isDomainRegistered = useSelector(state => state.search.currentSearchRegistered);
+  const crypto = useSelector(state => state.crypto);
+  const tempSearch = useSelector(state => state.search.tempSearchString);
+  const dispatch = useDispatch();
+  // tempSearchString
+  const { provider, wallet } = crypto;
+
+  
+  const deployDAO = async () => {    
+    const providerObj = getWalletProvider();
+    // setProvider(providerObj);
+    const walletObj = await getWallet(providerObj);
+    console.log('====================================');
+    console.log('in deploy dao', { providerObj, walletObj, tempSearch });
+    console.log('====================================');
+    const newDAOAddress = deployDomainDAO(walletObj, tempSearch, registryAddress);
+    console.log({ newDAOAddress });
+  }
+
+  return (
+    <div>
+      {
+        !isDomainRegistered
+        ?   <section id="no_dao_yet">
+              <div className="sidebar_inner_content_msg">
+                <h3>No DAO created for this URL.</h3>
+                <p>Creating a DAO will cost you 0.7ETH.</p>
+              </div>
+              <div className="button_box">
+                  <Button type="primary"  onClick={() => deployDAO()}>Create DAO</Button>
+                </div>
+            </section>
+        : <div>
+            <section id="user_has_not_invested">
+              <div className="sidebar_inner_content flex">
+                <div className="metrics">
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Total revenue generated by domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>123</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Total investments into domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>9382</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Return on Investment
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>2.6</h1>
+                      </div>
+                      <div className="metric_expression">
+                      %
+                      </div>
+                    </div>
+                  </div>
+                </div>
+                <div className="button_box">
+                  <Button type="primary">Sign Up to Invest</Button>
+                </div>
+              </div>
+            </section>
+
+            <section id="user_has_invested">
+              <div className="sidebar_inner_content flex">
+                <div className="metrics">
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Your revenue generated by domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>44</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Your investment into domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>9986</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Return on Investment
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>2.6</h1>
+                      </div>
+                      <div className="metric_expression">
+                      %
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div className="button_box">
+                  <Input suffix="ETH/$" />
+                  <Button type="primary">Add Investment</Button>
+                </div>
+              </div>
+            </section>
+          </div>
+      }
+    </div>
+  )
+}
+
+export default DeployDomain;
