+++ packages/desktop/src/components/finance/components/deployDomain/index.js
@@ -1,0 +1,3 @@
+import DeployDomain from './deployDomain';
+
+export default DeployDomain;
