+++ packages/desktop/src/components/finance/components/invested/index.js
@@ -1,0 +1,3 @@
+import Invested from './invested';
+
+export default Invested;
