+++ packages/desktop/src/components/finance/components/invested/invested.js
@@ -1,0 +1,70 @@
+import React, { Fragment } from 'react'
+import { useSelector } from 'react-redux';
+import { Button, Input } from 'antd';
+
+let registryAddress = "0x132efA3675cd66aA2780e01C095e2337188b0F6b";
+
+// NOTE: not integrated
+// Need another redux var to know if this user has invested in this domain
+const Invested = () => {
+  const isDomainRegistered = useSelector(state => state.search.currentSearchRegistered);
+
+  return (
+    <Fragment>
+      {
+        isDomainRegistered &&
+        <section id="user_has_invested">
+          <div className="sidebar_inner_content flex">
+            <div className="metrics">
+              <div className="metric_box">
+                <div className="metric_descr">
+                	Your revenue generated by domain
+                </div>
+                <div className="metric flex">
+                  <div className="metric_value">
+                    <h1>44</h1>
+                  </div>
+                  <div className="metric_expression">
+                    ETH/$
+                  </div>
+                </div>
+              </div>
+              <div className="metric_box">
+                <div className="metric_descr">
+                	Your investment into domain
+                </div>
+                <div className="metric flex">
+                  <div className="metric_value">
+                    <h1>9986</h1>
+                  </div>
+                  <div className="metric_expression">
+                    ETH/$
+                  </div>
+                </div>
+                </div>
+                <div className="metric_box">
+                	<div className="metric_descr">
+                		Return on Investment
+                	</div>
+                	<div className="metric flex">
+                    <div className="metric_value">
+                    	<h1>2.6</h1>
+                    </div>
+                    <div className="metric_expression">
+                    	%
+                    </div>
+                	</div>
+                </div>
+            	</div>
+							<div className="button_box">
+								<Input suffix="ETH/$" />
+								<Button type="primary">Add Investment</Button>
+							</div>
+            </div>
+        	</section>
+      }
+    </Fragment>
+  )
+}
+
+export default Invested;
