+++ packages/desktop/src/components/finance/components/notInvested/index.js
@@ -1,0 +1,3 @@
+import NotInvested from './notInvested';
+
+export default NotInvested;
