+++ packages/desktop/src/components/finance/components/notInvested/notInvested.js
@@ -1,0 +1,67 @@
+import React, { Fragment } from 'react'
+import { useSelector } from 'react-redux';
+import { Button, Input } from 'antd';
+
+// NOTE: not integrated
+// Need another redux var to know if this user has invested in this domain
+const NotInvested = () => {
+  const isDomainRegistered = useSelector(state => state.search.currentSearchRegistered);
+
+  return (
+    <Fragment>
+      {
+        isDomainRegistered &&
+            <section id="user_has_not_invested">
+              <div className="sidebar_inner_content flex">
+                <div className="metrics">
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Total revenue generated by domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>123</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Total investments into domain
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>9382</h1>
+                      </div>
+                      <div className="metric_expression">
+                      ETH/$
+                      </div>
+                    </div>
+                  </div>
+                  <div className="metric_box">
+                    <div className="metric_descr">
+                    Return on Investment
+                    </div>
+                    <div className="metric flex">
+                      <div className="metric_value">
+                        <h1>2.6</h1>
+                      </div>
+                      <div className="metric_expression">
+                      %
+                      </div>
+                    </div>
+                  </div>
+                </div>
+                <div className="button_box">
+                  <Button type="primary">Sign Up to Invest</Button>
+                </div>
+              </div>
+            </section>
+      }
+    </Fragment>
+  )
+}
+
+export default NotInvested;
