+++ packages/desktop/src/components/finance/finance.js
@@ -1,201 +1,15 @@
-import React, { useState } from 'react'
-import { useSelector, useDispatch } from 'react-redux';
-import { Button } from 'antd';
-import { Input } from 'antd';
-import { getWalletProvider, getWallet, deployDomainDAO } from '../../service/contracts/shared';
-import { toggleSider, siteNavigation  } from '../../redux/actions/actionCreators/siteNavigation';
-
-let registryAddress = "0x132efA3675cd66aA2780e01C095e2337188b0F6b";
+import React, { Fragment } from 'react'
+import DeployDomain from './components/deployDomain';
+import Invested from './components/invested';
+import NotInvested from './components/notInvested';
 
 const Finance = () => {
-  // need to know if this current domain has been registered
-  // if it is deployed and if you have invested
-  // render out diff stuff depending on the above
-  const isDomainRegistered = useSelector(state => state.search.currentSearchRegistered);
-  const crypto = useSelector(state => state.crypto);
-  const tempSearch = useSelector(state => state.search.tempSearchString);
-  const dispatch = useDispatch();
-  // tempSearchString
-  const { provider, wallet } = crypto;
-
-  
-  const deployDAO = async () => {    
-    const providerObj = getWalletProvider();
-    // setProvider(providerObj);
-    const walletObj = await getWallet(providerObj);
-    console.log('====================================');
-    console.log('in deploy dao', { providerObj, walletObj, tempSearch });
-    console.log('====================================');
-    const newDAOAddress = deployDomainDAO(walletObj, tempSearch, registryAddress);
-    console.log({ newDAOAddress });
-  }
-
-  const closeMenu = () => {
-    dispatch(toggleSider(false))
-    dispatch(siteNavigation(10))
-  }
-
-  return (
-    // <div className="sidebar_inner expanded flex">
-    <div>
-      {/* <div className="sidebar_inner_title flex">
-        <div className="sidebar_inner_title_icon">
-          <svg x="0px" y="0px" viewBox="0 0 38.1 36.8">
-                <path className="st3" d="M24.3,9.2c-0.3,0-0.5,0.2-0.5,0.5s0.2,0.5,0.5,0.5h2.1l-7.3,7.3L16,14.4l-6.3,6.3c-0.2,0.2-0.2,0.5,0,0.7
-                    c0.1,0.1,0.2,0.1,0.4,0.1s0.3,0,0.4-0.1l5.6-5.6l3.1,3.1l8-8V13c0,0.3,0.2,0.5,0.5,0.5s0.5-0.2,0.5-0.5V9.2H24.3z"/>
-                <path className="st3" d="M10.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-2.5c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v2.5
-                    C10.6,27.4,10.4,27.6,10.1,27.6z"/>
-                <path className="st3" d="M13.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-4.6c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v4.6
-                    C13.6,27.4,13.4,27.6,13.1,27.6z"/>
-                <path className="st3" d="M16.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-6.5c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v6.5
-                    C16.6,27.4,16.4,27.6,16.1,27.6z"/>
-                <path className="st3" d="M19.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-4.5c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v4.5
-                    C19.6,27.4,19.4,27.6,19.1,27.6z"/>
-                <path className="st3" d="M22.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-6.7c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v6.7
-                    C22.6,27.4,22.4,27.6,22.1,27.6z"/>
-                <path className="st3" d="M25.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-9c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v9C25.6,27.4,25.4,27.6,25.1,27.6z
-                    "/>
-                <path className="st3" d="M28.1,27.6c-0.3,0-0.5-0.2-0.5-0.5v-10c0-0.3,0.2-0.5,0.5-0.5s0.5,0.2,0.5,0.5v10
-                    C28.6,27.4,28.4,27.6,28.1,27.6z"/>
-            </svg>
-        </div>
-        <div className="spacer"></div>
-        <Button size="large" shape="circle" onClick={() => closeMenu()}>
-          <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" viewBox="0 0 38.1 36.8">
-            <g>
-              <path className="st0" d="M26.5,26.9c-0.3,0-0.5-0.1-0.7-0.3l-15-15c-0.4-0.4-0.4-1,0-1.4s1-0.4,1.4,0l15,15c0.4,0.4,0.4,1,0,1.4
-                C27,26.8,26.8,26.9,26.5,26.9z"/>
-            </g>
-            <g>
-              <path className="st0" d="M11.6,26.9c-0.3,0-0.5-0.1-0.7-0.3c-0.4-0.4-0.4-1,0-1.4l15-15c0.4-0.4,1-0.4,1.4,0s0.4,1,0,1.4l-15,15
-                C12.1,26.8,11.8,26.9,11.6,26.9z"/>
-            </g>
-          </svg>
-        </Button>
-      </div> */}
-
-      {
-        !isDomainRegistered
-        ?   <section id="no_dao_yet">
-              <div className="sidebar_inner_content_msg">
-                <h3>No DAO created for this URL.</h3>
-                <p>Creating a DAO will cost you 0.7ETH.</p>
-              </div>
-              <div className="button_box">
-                  <Button type="primary"  onClick={() => deployDAO()}>Create DAO</Button>
-                </div>
-            </section>
-        : <div>
-            <section id="user_has_not_invested">
-              <div className="sidebar_inner_content flex">
-                <div className="metrics">
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Total revenue generated by domain
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>123</h1>
-                      </div>
-                      <div className="metric_expression">
-                      ETH/$
-                      </div>
-                    </div>
-                  </div>
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Total investments into domain
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>9382</h1>
-                      </div>
-                      <div className="metric_expression">
-                      ETH/$
-                      </div>
-                    </div>
-                  </div>
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Return on Investment
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>2.6</h1>
-                      </div>
-                      <div className="metric_expression">
-                      %
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div className="button_box">
-                  <Button type="primary">Sign Up to Invest</Button>
-                </div>
-              </div>
-            </section>
-
-            <section id="user_has_invested">
-              <div className="sidebar_inner_content flex">
-                <div className="metrics">
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Your revenue generated by domain
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>44</h1>
-                      </div>
-                      <div className="metric_expression">
-                      ETH/$
-                      </div>
-                    </div>
-                  </div>
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Your investment into domain
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>9986</h1>
-                      </div>
-                      <div className="metric_expression">
-                      ETH/$
-                      </div>
-                    </div>
-                  </div>
-                  <div className="metric_box">
-                    <div className="metric_descr">
-                    Return on Investment
-                    </div>
-                    <div className="metric flex">
-                      <div className="metric_value">
-                        <h1>2.6</h1>
-                      </div>
-                      <div className="metric_expression">
-                      %
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-                <div className="button_box">
-                  <Input suffix="ETH/$" />
-                  <Button type="primary">Add Investment</Button>
-                </div>
-              </div>
-            </section>
-          </div>
-      }
-
-
-    
-
-
-
-
-
-    </div>
+  return (  
+    <Fragment>
+      <DeployDomain />
+      <Invested />
+      <NotInvested />
+    </Fragment>
   )
 }
